# Shortcut `List<T>` for `Enumerable.Any()`